`altitude_azimuth_mesh()` moved to `_common_new`
Generator function `altitude_azimuth_mesh()` yielding more-or-less
regular grid of altitude/azimuth coordinates is now available in the
`_common_new` module.
