+++ extras/_kesi_dipolar.py
@@ -1,14 +1,4 @@
-from _common_new import FourSphereModel, cv
-
-def altitude_azimuth_mesh(base, step, alternate=True):
-    for i, altitude in enumerate(np.linspace(base,
-                                             np.pi / 2,
-                                             int(round((np.pi / 2 - base) / step) + 1))):
-        for azimuth in (np.linspace(-1, 1,
-                                    int(round(np.cos(altitude) * 2 * np.pi / step)) + 1,
-                                    endpoint=False)
-                        + (0.5 * (-1) ** i if alternate else 1)) * np.pi:
-            yield altitude, azimuth
+from _common_new import FourSphereModel, cv, altitude_azimuth_mesh
 
 
 class DipolarSourcesFactory(object):
